Base foraging on circle rather than square grid
Using a square grid for foraging results in varying movement
length, because for NSEW movement we're using a straight line,
and for movement toward the corners we're using a diagonal.
Using a circle allows us to always use the radius of the circle,
so each hop to a new location is the same length as every
other hop.