Fix #1: archons clustering near origin
After many hours of pointless fooling around with random number
generators, I've discovered that the problem was that I was calling
floor() on candidate foraging targets. The idea was simply to get rid
of decimals in order to make == work. I had no idea that truncating
a decimal would have such a profound effect. I ended up not
using == anyway.